+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ### Added
 
 - Added ADC helper functions to read more intuitive values (#22) - @HarkonenBade
+- Added interrupt enabling/disabling support to USART ports
 
 ### Changed